feat: implement chapter11 (auth, storage, upload)

+++ src/app/admin/posts/[id]/page.tsx
@@ -6,7 +6,9 @@ import PostForm from "@/app/admin/_components/PostForm";
 import type { PostFormValues } from "@/app/admin/_components/PostForm";
 import { BackButton, DeleteButton, UpdateButton } from "@/app/admin/_components/Button";
 import type { CategoriesIndexResponse } from "@/app/api/admin/categories/route";
-import type { PostShowResponse } from "@/app/api/posts/[id]/route";
+import type { PostShowResponse } from "@/app/api/admin/posts/[id]/route";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
+import { useFetch } from "@/app/_hooks/useFetch";
 
 type Props = {
   params: {
@@ -14,100 +16,60 @@ type Props = {
   };
 };
 
-type Category = PostShowResponse["post"]["postCategories"][number]["category"]
-
 export default function AdminPostsEditPage({params}: Props) {
   const router = useRouter();
   const [title, setTitle] = useState("");
   const [content, setContent] = useState("");
-  const [thumbnailUrl, setThumbnailUrl] = useState("");
+  const [thumbnailImageKey, setThumbnailImageKey] = useState("");
   const [selectedCategoryIds, setSelectedCategoryIds] = useState<number[]>([]);
-  const [categories, setCategories] = useState<Category[]>([]);
-  const [, setIsSubmitting] = useState(false);
-  const [isLoading, setIsLoading] = useState(true);
+  const [isSubmitting, setIsSubmitting] = useState(false);
   const [errorMessage, setErrorMessage] = useState("");
-  const [, setTitleErrorMessage] = useState("");
-  const [, setContentErrorMessage] = useState("");
-  const [, setThumbnailUrlErrorMessage] = useState("");
+  const { token } = useSupabaseSession();
+  const {
+    data: postData,
+    error: postError,
+    isLoading: isPostLoading,
+  } = useFetch<PostShowResponse>(
+    `/api/admin/posts/${params.id}`
+  );
+  const post = postData?.post;
+  const {
+    data: categoriesData,
+    error: categoriesError,
+    isLoading: isCategoriesLoading,
+  } = useFetch<CategoriesIndexResponse>(
+    "/api/admin/categories"
+  )
+  const categories = categoriesData?.categories;  
 
-  // 記事の情報取得
   useEffect(() => {
-    const fetchPost = async () => {
-      try {
-        const res = await fetch(`/api/admin/posts/${params.id}`);
+    if (!post) return;
 
-        if (!res.ok) {
-          throw new Error("記事の情報の取得に失敗しました");
-        }
-
-        const data = await res.json();
-        setTitle(data.post.title);
-        setContent(data.post.content);
-        setThumbnailUrl(data.post.thumbnailUrl);
-        //中間テーブル
-        setSelectedCategoryIds(
-          data.post.postCategories.map(
-            (postCategory: { category: { id: number } }) => postCategory.category.id
-          )
-        );
-
-        console.log(data);
-        
-      } catch (error) {
-        console.error(error);
-        setErrorMessage("記事の情報の取得に失敗しました");
-      } finally {
-        setIsLoading(false);
-      }
-    }
-    fetchPost();
-  }, [params.id]);
-
-  // カテゴリーの取得
-  useEffect(() => {
-    const fetchCategories = async () => {
-        const res = await fetch(`/api/admin/categories`);
-        const data: CategoriesIndexResponse = await res.json();
-        if (!res.ok) {
-          throw new Error("カテゴリの取得に失敗しました");
-        }
-        setCategories(data.categories);
-      };
-    fetchCategories();
-  }, []);
+    setTitle(post.title);
+    setContent(post.content);
+    setThumbnailImageKey(post.thumbnailImageKey);
+    setSelectedCategoryIds(
+      post.postCategories.map((postCategory) => postCategory.category.id)
+    );
+  },[post]);
 
   const handleUpdate = async (values: PostFormValues) => {
-    if (!values.title.trim()) {
-      setTitleErrorMessage("タイトルを入力してください");
-      return;
+    if (!token) {
+      throw new Error('ログイン情報が取得できません');
     }
-
-    if (!values.content.trim()) {
-      setContentErrorMessage("記事内容を入力してください");
-      return;
-    }
-
-    if (!values.thumbnailUrl.trim()) {
-      setThumbnailUrlErrorMessage("サムネイルURLを入力してください");
-      return;
-    }
-
 
     try {
+      setIsSubmitting(true);
+      setErrorMessage("");
+
       const res = await fetch(`/api/admin/posts/${params.id}`, {
         method: "PUT",
         headers: {
           "Content-Type": "application/json",
+          Authorization: token,
         },
-        body: JSON.stringify({
-          title: values.title,
-          content: values.content,
-          thumbnailUrl: values.thumbnailUrl,
-          categoryIds: values.categoryIds,
-        }),
+        body: JSON.stringify(values),
       });
-
-      console.log("status", res)
 
       if (!res.ok) {
         throw new Error("記事の更新に失敗しました");
@@ -126,12 +88,21 @@ export default function AdminPostsEditPage({params}: Props) {
     if (!confirm("本当にこの記事を削除しますか？")) {
       return;
     }
-    
+
+    if (!token) {
+      setErrorMessage("ログイン情報が取得できません");
+      return;
+    }
+
     try {
       setIsSubmitting(true);
       setErrorMessage("");
+      
       const res = await fetch(`/api/admin/posts/${params.id}`, {
-        method: "DELETE"
+        method: "DELETE",
+        headers: {
+          Authorization: token,
+        },
       });
 
       if (!res.ok) {
@@ -147,22 +118,28 @@ export default function AdminPostsEditPage({params}: Props) {
     }
   }
 
-  if (isLoading) {
+  if (isPostLoading || isCategoriesLoading) {
     return <p>Loading...</p>;
+  }
+
+  if (postError || categoriesError) {
+    return <p>データの取得に失敗しました</p>
   }
 
   return (
     <div>
-      <h1 className="text-2xl font-bold mb-2">記事編集</h1>
+      <div className="mb-6 flex items-center justify-between">
+        <h1 className="text-2xl font-bold py-[4px]">記事編集</h1>
+      </div>
       <PostForm
         formId="post-edit-form"
         initialValues={{
           title,
           content,
-          thumbnailUrl,
+          thumbnailImageKey,
           categoryIds: selectedCategoryIds,
         }}
-        categories={categories}
+        categories={categories ?? []}
         onSubmit={handleUpdate}
       />
 
@@ -171,9 +148,8 @@ export default function AdminPostsEditPage({params}: Props) {
       )}  
 
       <div className="mt-4 flex gap-3">
-        <UpdateButton form="post-edit-form" />
-        <DeleteButton onClick={handleDelete} />
-
+        <UpdateButton form="post-edit-form" disabled={isSubmitting}/>
+        <DeleteButton onClick={handleDelete} disabled={isSubmitting}/>
         <BackButton href="/admin/posts" />
       </div>
     </div>
